Split view_state.py into separate files
