+++ src/view/ui_element.py
@@ -1,0 +1,4 @@
+from view_element import ViewElement
+
+class UIElement(ViewElement):
+    pass
